pyproject.toml
[tool.poetry]
name = "beaverhabits"
version = "0.1.0"
description = ""
authors = ["Chanjian Zhu <daya0576@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
nicegui = "^1.4.18"
pydantic-settings = "^2.2.1"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
aiosqlite = "^0.20.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
icecream = "^2.1.3"
autopep8 = "^2.0.4"
pytest-selenium = "^4.1.0"
pytest-asyncio = "^0.23.5.post1"
webdriver-manager = "^4.0.1"
selenium = "^4.18.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "beaverhabits.routes:main"
